[ADD] add time and cpu in proc_info struct in top sys_call

+++ kernel/proc.h
@@ -114,7 +114,7 @@ struct proc_info {
     int pid;
     int ppid;
     long time;
-    int cpu;
+    float cpu;
     enum procstate state;
 };
