+++ user/top.h
@@ -5,7 +5,7 @@ struct proc_info {
     int pid;
     int ppid;
     long time;
-    int cpu;
+    float cpu;
     enum procstate state;
 };
